Update 345 CSQ OI - Spring 23.md
Here I adjusted some dates that were wrong (i.e. "this Fall 2022 semester" ->> "this Spring 2023 semester"). One thing to mention is mid semester flight awards was scheduled on LLAB 6 last semester but it is not on the master overview at all. I went in and penciled the date for this LLAB 6 but we can adjust accordingly